google-cloud-run/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
firebase-admin
google-cloud-firestore